move test helper functions to there own module
